Render python_version option only if both python versions are present
If both versions are present, the --with-both-python option is passed
to inv agent.build, which will generate the correct configuration.

The generation is probably suboptimal: we generate datadog.yaml 3 times,
because we first generate every config (so one for py3only, another one
for py2py3), and then we specifically generate the config we were asked
to generate (either one or the other).
This was done because with Go regexes, there's no easy way to exclude
lines that contain a particular word, and I didn't want to list all other
generated config files (that'd mean having to add it here when we add a
new config template).